evaluate/requirements.txt
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pandas>=2.2,<3.0
numpy>=1.26,<3.0
scipy>=1.13
statsmodels>=0.14
python-dotenv>=1.0
pyyaml>=6.0
tabulate>=0.9
rich>=13.7
scikit-posthocs>=0.9
